🔧 ✅ 🔥 Major changes to the folder strucure of the graphql schema and resolvers
- Removed some unnessary configutaion setup files like .babel.json,
  .v8flags and shrinkwrap.yaml :neutral_face:
- Simplified some resolvers to have their own folders for easier
  refactoring and testing :relaxed:
