+++ package.json
@@ -10,7 +10,7 @@
 	"scripts": {
 		"test": "jest",
 		"dev": "nodemon src/index.js --exec babel-node -e js",
-		"clean": "rm -rf build && mkdir build",
+		"clean": "rm -rf build && rm -rf coverage && mkdir build",
 		"build-babel": "babel -d ./build ./src -s",
 		"build": "npm run clean && npm run build-babel",
 		"start": "npm run build && node ./build/index.js"
