+++ src/resolvers/queries/drivers.js
@@ -1,0 +1,6 @@
+// ANCHOR this will return all drivers
+import Drivers from '../../models/Drivers';
+
+export default function drivers () {
+    return Drivers.find().catch((error) => console.log(error));
+}
