+++ src/resolvers/utils.js
@@ -1,0 +1,4 @@
+// ANCHOR Here we store utility functions that are share by all
+
+// NOTE Parses the input and returns an object
+export const parceArgs = (input) => JSON.parse(JSON.stringify(input)); // TODO Test function
